fix: align with guidance on input for text fields
This commit removes the `maxlength` attribute from the TextField
component which brings it in-line with the design system guidance on
how to ask for text answers.

Note that because the TextField component uses `joi.string()` as its
base schema and implements the `max` property validation on the backend
there is no additional validation that has to be applied, this change
will only stop the browser from restricting input after the specified
`max` propety has been reached.

Detail on this guidance is here:
https://design-system.service.gov.uk/components/text-input#avoid-restricting-the-length-of-a-users-input