fix: respect the NumberField precision option
It looks like the `precision` option that can be specified for the
NumberField component is not appropriately applied the view model.

A previous refactor looks like it split out the `schema` (which now
refers to the Joi validation schema) and `schemaOptions` which can
be provided to the component to configure various properties.

As the code still refers to `this.schema` it doesn't matter what property
is set by the component constructur, it will always only default to setting
`0.1` as the `step` HTML attribute for any truthy value.

Using `this.schemaOptions` should allow higher numbers to set a smaller
granularity as it looks like was expected.

Add coverage that should break with the existing code and is addressed
by this change.